+++ src/worker/util/changes.football.ts
@@ -73,6 +73,11 @@ const changes = [
 		msg:
 			'Added a new "GM History" page in the Team menu, where you can see your performance across all the different franchises you managed in the past. <a href="https://basketball-gm.com/blog/2020/08/gm-history/">More details</a>',
 	},
+	{
+		date: "2020-08-22",
+		msg:
+			'New God Mode option: Spectator Mode, where the AI controls all teams and you just watch. It\'s like auto play, but you can proceed at your own pace. <a href="https://basketball-gm.com/blog/2020/08/spectator-mode/">More details</a>',
+	},
 ];
 
 export default changes;